Test Model
Test model got some love. I'm not an artist by any stretch of the imagination but I need something to test with that I own the rights to. The test_model has a skeleton and 4 cubes that are weighted to the skeleton. I have also prepared textures and offset the UVs in preparation of the MaterialsPane.