Prototype
This is the original version of AWOC for Godot 4.3. I was learning Godot and prototyping ideas. In a race to prove the concept, I took a lot of shortcuts. This version works for me without errors, but it is difficult to maintain because I gave very little thought to structure. A rewrite is currently underway.