Rewrite Begin
This is where I am currently with the rewrite. The foundation still needs some work before I begin adding more tabs. In particular, the base classes aren't doing as much as they should be doing. Additionally, the mesh data is no longer being stored as individual custom resources on disk. Instead, the meshes are saved as array meshes in the AWOC resource. The meshes are stored correctly in 4.4 but rebuilding them seems to be broken. So I am continuing to develop in 4.3 until either the array mesh problem gets fixed or I finish the rewrite and have time to figure out another solution.