Big boy britches
AWOC has moved to the main viewport. This will allow the main editor interface to be on the left and the preview/details window on the right. Instead of having a HideSlotTab for each slot and an OverlayTab for each Material, these tabs will be shown in the details area.