Mesh Preview
The logic for the eye toggle button in mesh_row_management was roughly sketched in. Now, it works as intended. If all eyes are open, waiting to be clicked on and closed, the Preview area is hidden. But if one or more eyes are closed, the Preview area is shown. This makes more sense when you see it in action.